Added more entities. Working on the services and postman

+++ src/main/java/com/carparkspring/demo/repository/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<AppUser, Integer> {
+public interface UserRepository extends JpaRepository<AppUser, Long> {
     Optional<AppUser> findByEmailId(String email);
 }
